split the state module into several source files.
